requirements.txt
Django>=4.2
djangorestframework
python-dotenv
gspread
oauth2client
stripe
djangorestframework-simplejwt